+++ src/jvmTest/kotlin/URLTest.kt
@@ -1,52 +1,48 @@
 package garden.ephemeral.clipninja
 
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import org.junit.jupiter.api.Test
+import io.kotest.core.spec.style.FreeSpec
+import io.kotest.matchers.shouldBe
 
-class URLTest {
-    @Test
-    fun `parsing no query`() {
-        assertThat(URL.fromString("https://example.com/"))
-            .isEqualTo(URL("https", "example.com", "/", null))
-    }
+class URLTest : FreeSpec({
+    "fromString" - {
+        "parsing no query" {
+            URL.fromString("https://example.com/")
+                .shouldBe(URL("https", "example.com", "/", null))
+        }
 
-    @Test
-    fun `formatting no query`() {
-        assertThat(URL("https", "example.com", "/", null).toString())
-            .isEqualTo("https://example.com/")
-    }
+        "parsing one query param" {
+            URL.fromString("https://example.com/search?q=blah")
+                .shouldBe(URL("https", "example.com", "/search", mapOf("q" to "blah")))
+        }
 
-    @Test
-    fun `parsing one query param`() {
-        assertThat(URL.fromString("https://example.com/search?q=blah"))
-            .isEqualTo(URL("https", "example.com", "/search", mapOf("q" to "blah")))
-    }
-
-    @Test
-    fun `formatting one query param`() {
-        assertThat(URL("https", "example.com", "/search", mapOf("q" to "blah")).toString())
-            .isEqualTo("https://example.com/search?q=blah")
-    }
-
-    @Test
-    fun `parsing two query params`() {
-        assertThat(URL.fromString("https://example.com/search?q=blah&utm_source=somewhere"))
-            .isEqualTo(
-                URL(
-                    "https", "example.com", "/search",
-                    mapOf("q" to "blah", "utm_source" to "somewhere")
+        "parsing two query params" {
+            URL.fromString("https://example.com/search?q=blah&utm_source=somewhere")
+                .shouldBe(
+                    URL(
+                        "https", "example.com", "/search",
+                        mapOf("q" to "blah", "utm_source" to "somewhere")
+                    )
                 )
-            )
+        }
     }
 
-    @Test
-    fun `formatting two query params`() {
-        assertThat(
-            URL(
-                "https", "example.com", "/search",
-                mapOf("q" to "blah", "utm_source" to "somewhere")
-            ).toString()
-        ).isEqualTo("https://example.com/search?q=blah&utm_source=somewhere")
+    "toString" - {
+        "formatting no query" {
+            URL("https", "example.com", "/", null)
+                .toString()
+                .shouldBe("https://example.com/")
+        }
+
+        "formatting one query param" {
+            URL("https", "example.com", "/search", mapOf("q" to "blah"))
+                .toString()
+                .shouldBe("https://example.com/search?q=blah")
+        }
+
+        "formatting two query params" {
+            URL("https", "example.com", "/search", mapOf("q" to "blah", "utm_source" to "somewhere"))
+                .toString()
+                .shouldBe("https://example.com/search?q=blah&utm_source=somewhere")
+        }
     }
-}
+})