+++ Changes.md
@@ -2,6 +2,7 @@
 - Added a second level of safety switch for showing all strings
 - Renamed application to avoid stepping on someone else's name
 - Copy button to copy previous clipboard content back to the clipboard
+- Fixed removal of tracking parameters for YouTube URLs
 
 # 1.0.2
 - Slimmed down the installer by including fewer icons